[Ghalnas-33] feat : back end logic for log displaying

+++ src/model/Position.java
@@ -34,7 +34,7 @@ public class Position
     @Override
     public String toString()
     {
-        return "(" + x + "," + y + ")";
+        return "(" + (int)x + "," + (int)y + ")";
     }
 
     @Override
